split chip files into readonly and interactive
