+++ src/components/DataDisplay/Chip/Chip.tsx
@@ -1,17 +1,15 @@
 import {
   Children as ReactChildren,
-  forwardRef,
   Fragment,
-  HTMLAttributes,
   isValidElement,
   ReactElement,
   ReactNode,
 } from 'react';
 
-import { Icon } from '@equinor/eds-core-react';
 import { IconData } from '@equinor/eds-icons';
 
-import { InteractiveChipStyle, ReadOnlyChipStyle } from './Chip.styles';
+import { InteractiveChip, InteractiveChipProps } from './InteractiveChip';
+import { ReadOnlyChip, ReadOnlyChipProps } from './ReadOnlyChip';
 
 // Base type for common properties
 export interface BaseChipProps {
@@ -29,99 +27,7 @@ export interface BaseChipProps {
   leadingIconData?: IconData;
 }
 
-// Type for ReadOnlyChipStyle (div)
-export type ReadOnlyChipProps = BaseChipProps & HTMLAttributes<HTMLDivElement>;
-
-// Type for InteractiveChipStyle (button)
-export type InteractiveChipProps = BaseChipProps &
-  HTMLAttributes<HTMLButtonElement>;
-
 type ReactElementWithChildren = ReactElement<{ children?: ReactNode }>;
-
-// Define the ReadOnlyChip component
-const ReadOnlyChip = forwardRef<HTMLDivElement, ReadOnlyChipProps>(
-  (props, ref) => {
-    const { children, disabled, leadingIconData, ...otherReadOnlyProps } =
-      props;
-
-    // ReadOnlyChip only needs to handle the div related props and children,
-    return (
-      <ReadOnlyChipStyle
-        {...otherReadOnlyProps}
-        className={disabled ? 'disabled' : ''}
-        ref={ref}
-      >
-        <div className="content">
-          {leadingIconData && (
-            <div className="leading">
-              <Icon data={leadingIconData} size={16} />
-            </div>
-          )}
-          {children}
-        </div>
-      </ReadOnlyChipStyle>
-    );
-  }
-);
-ReadOnlyChip.displayName = 'ReadOnlyChip';
-
-// Define the InteractiveChip component
-const InteractiveChip = forwardRef<HTMLButtonElement, InteractiveChipProps>(
-  (props, ref) => {
-    const {
-      children,
-      onDelete,
-      disabled = false,
-      variant,
-      onClick,
-      leadingIconData,
-      ...otherInteractiveProps
-    } = props;
-
-    // deletable and clickable logic can be handled within this component
-    const deletable = onDelete !== undefined;
-    const clickable = !disabled && props.onClick !== undefined;
-    const handleDelete = disabled ? undefined : onDelete;
-    const handleClick = disabled ? undefined : onClick;
-
-    const handleKeyPress = (
-      event:
-        | React.KeyboardEvent<HTMLButtonElement>
-        | React.MouseEvent<HTMLButtonElement>
-    ) => {
-      const { key } = event as React.KeyboardEvent<HTMLButtonElement>;
-      if (key === 'Enter') {
-        if (deletable) {
-          handleDelete!(event as unknown); // using non-null assertion operator since we know deletable is true
-        } else if (clickable) {
-          handleClick!(event as unknown); // using non-null assertion operator since we know clickable is true
-        }
-      }
-    };
-
-    return (
-      <InteractiveChipStyle
-        {...otherInteractiveProps}
-        ref={ref}
-        disabled={disabled}
-        variant={variant}
-        onClick={deletable ? onDelete : onClick}
-        onKeyDown={handleKeyPress}
-      >
-        <div className="content">
-          {leadingIconData && (
-            <div className="leading">
-              <Icon role={'img'} data={leadingIconData} size={16} />
-            </div>
-          )}
-          {children}
-          {deletable && <Icon name="close" title="close" size={16} />}
-        </div>
-      </InteractiveChipStyle>
-    );
-  }
-);
-InteractiveChip.displayName = 'InteractiveChip';
 
 export const Chip = ({
   children,
